Remove bluebird
Bluebird was only used for `promisifyAll` on the bag reader to easily add promise versions of a few internal methods.  Bluebird increases the bundle size of rosbag when used in the browser substantially, so I opted to remove it as a dependency and hand-roll the few places where promise versions of functions are called.