Resolve unqualified types to fully qualified types
To make it easier to figure out which definitions refer to which types.

I also made the output of `getTypes` simple objects, instead of class
instances, as the latter may imply that there are any functions or so
callable on the instances, or that they’re not serialisable, and neither
is actually true.

Also actually export `getTypes`. :)

Test plan: actually tested this out in a real application, and wrote a
unit test.