improve error handling for missing and empty files (#27)
Fix some poor error propagation reported in #25. Adds a test for reading a nonexistent file as well as a completely empty file.

I had also previously noticed that a valid but _empty_ bag file (no chunks/connections) caused an error when we tried to access `connections[connectionCount - 1].end`, so I added a check for `connectionCount === 0`.