Improve TimeUtil.add
Previously calling `TimeUtil.add` with negative values in either sec or nsec resulted in undefined-like behavior, producing negative values in the result without taking into consideration rounding of nsecs to secs.  This change allows for adding a negative time (duration) value to an existing time to produce a new adjusted time with the negative values properly subtracted.  If the resulting time results in a negative value, we now throw an exception instead of returning an invalid time value.

Test plan: updated unit tests