Add “freeze” option for building immutable messages (#60)
This way we can prevent our code or libraries from mutating raw
messages.

Initially I tried to do this on existing messages using the
`simple-deep-freeze` library, but that only works on “simple” objects
(without a custom prototype), but we want to keep custom prototypes
since they look better in the browser console, and they are necessary if
in the future we build lazy parsing. So just freezing them within
rosbag.js is better since that way we have more control over exactly how
the freezing works.

Test plan: added a bunch of tests.