Do not inline everything in generated parse code (#68)
Instead of inlining every field for every submessage recursively, we
should instead use functions for distinct message types.

For types like

    A { B b1, b2 };
    B { C c1, c2 };
    ...
    N { int dummy };

the size of the generated code is O(N) rather than O(2^n), and the size
of the largest function body is O(1) instead of O(2^n). This can make a
large performance difference for very large message definition DAGs in
Chrome where functions with code size larger than 128KB are are not
optimised. This results in a large speedup for some common webviz
use-cases.

(One possible downside, the maximum stack depth is now linear, not
constant. Probably not an issue in practice?)

The total number of function calls is not increased by this change, and
the the removal of the `this.field_X = undefined;` logic doesn't hurt
the object layout -- this results in a measured playback speedup of
about 5% in webviz testing for me.

As a nice upside, this represents a nice simplification of both the
generating and generated code.

Test plan: No change in functionality. Covered by existing tests.