Add message writer API
Add a message writer API that allows writing a message from a JS
object to ROS binary. This does not currently allow bag writing, and
I don't currently have plans to build or use that feature, so this
is mostly standalone for writing individual messages. It is mostly
translated from the existing MessageReader class.

Before writing the message we need to know the binary length of the
message, so we can create the buffer with that size. This currently
requires traversing the message twice, which is likely not very
efficient. I'm going to start with this approach and measure how
performant I need to make this before making further improvements.

Test plan: included all relevant tests from the MessageReader, just
converted to test writing instead.